evals/pyproject.toml
[project]
name = "evals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Quinn Dougherty", email = "quinn.dougherty.forall@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.4",
    "pydantic-ai>=1.19.0",
    "datasets>=4.4.1",
    "gitpython>=3.1.40",
    "inspect-ai>=0.3.146",
    "typer>=0.20.0",
    "anthropic>=0.73.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
agent = "evals:main"
mine = "evals.dalek_history.cli:main"

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"
